% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tonality.R
\name{document_scores}
\alias{document_scores}
\alias{sentence_scores}
\alias{ngram_scores}
\title{Extract content from a Gavagai tonality object}
\usage{
document_scores(x)

sentence_scores(x)

ngram_scores(x)
}
\arguments{
\item{x}{a \code{gavagai_tonality} object.}
}
\description{
Extract content from a Gavagai tonality object
}
