% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tonality.R
\name{extract_tonality_ngrams}
\alias{extract_tonality_ngrams}
\title{Extract ngram with scores}
\usage{
extract_tonality_ngrams(x)
}
\arguments{
\item{x}{the text slot in a API result.}
}
\description{
Extract ngram with scores
}
