% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tonality.R
\name{extract_tonality_score_by_doc}
\alias{extract_tonality_score_by_doc}
\title{Extract document level sentiments and scores}
\usage{
extract_tonality_score_by_doc(x)
}
\arguments{
\item{x}{the text slot in a API result.}
}
\description{
Extract document level sentiments and scores
}
