% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tonality.R
\name{extract_tonality_score_by_sentence}
\alias{extract_tonality_score_by_sentence}
\title{Extract sentence level sentiments and scores}
\usage{
extract_tonality_score_by_sentence(x)
}
\arguments{
\item{x}{the text slot in a API result.}
}
\description{
Extract sentence level sentiments and scores
}
