% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_languages.R
\name{supported_languages}
\alias{supported_languages}
\title{Extract supported languages from the Gavagai API}
\usage{
supported_languages(api_key)
}
\arguments{
\item{api_key}{API key to use to connect to the API.}
}
\description{
Extract supported languages from the Gavagai API
}
