% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tonality.R
\name{tonality}
\alias{tonality}
\title{Get tonality of text from Gavagai API}
\usage{
tonality(texts, api_key, language, tonality = NULL)
}
\arguments{
\item{texts}{The texts to extract tonality from.}

\item{api_key}{API key to use to connect to the API.}

\item{language}{The language code in ISO 639-1}

\item{tonality}{What tonality should be returned. \code{NULL} is all.

texts <- c("I love you!", "Indeterminacy of reference refers to the interpretation of words or phrases in isolation, and Quine's thesis is that no unique interpretation is possible, because a 'radical interpreter' has no way of telling which of many possible meanings the speaker has in mind. Quine uses the example of the word 'gavagai' uttered by a native speaker of the unknown language Arunta upon seeing a rabbit.")
language <- "en"}
}
\description{
Get tonality of text from Gavagai API
}
\details{
See \code{https://developer.gavagai.se/}.
}
