% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tonality.R
\name{tonality_document}
\alias{tonality_document}
\alias{tonality_sentence}
\alias{tonality_ngram}
\title{Extract content from a Gavagai tonality object}
\usage{
tonality_document(x)

tonality_sentence(x)

tonality_ngram(x)
}
\arguments{
\item{x}{a \code{gavagai_tonality} object.}
}
\description{
Extract content from a Gavagai tonality object
}
